refactor: guidechat 패키지 분리
