+++ backend/src/main/java/kr/ac/hs/RandomTrip/trip/service/TourService.java
@@ -4,8 +4,8 @@ import kr.ac.hs.RandomTrip.auth.domain.Member;
 import kr.ac.hs.RandomTrip.auth.repository.MemberRepository;
 import kr.ac.hs.RandomTrip.trip.domain.Destination;
 import kr.ac.hs.RandomTrip.trip.domain.Visit;
-import kr.ac.hs.RandomTrip.trip.dto.GuideResponse;
-import kr.ac.hs.RandomTrip.trip.llm.GeminiGuideGenerator;
+import kr.ac.hs.RandomTrip.guidechat.dto.GuideResponse;
+import kr.ac.hs.RandomTrip.guidechat.llm.GeminiGuideGenerator;
 import kr.ac.hs.RandomTrip.trip.repository.DestinationRepository;
 import kr.ac.hs.RandomTrip.itinerary.domain.ItineraryItem;
 import kr.ac.hs.RandomTrip.itinerary.repository.ItineraryItemRepository;
